Ajout de fonctionnalités via les modules os et path

+++ 00_Livrables/scraper_categories.py
@@ -4,6 +4,22 @@
 
 import parser_url
 import csv
+import os
+from pathlib import Path
+
+directory = str(Path(__file__).resolve().parent)
+
+user_auth_rm_csv = ""
+
+if os.path.exists(directory + "/output_scraper.csv") :
+    while user_auth_rm_csv != "Y" and user_auth_rm_csv != "n" :
+        user_auth_rm_csv = input("Le fichier output_scraper.csv existe déja. Souhaitez-vous l'effacer et relancer l'import de données? [Y/n] : ")
+        if user_auth_rm_csv != "Y" and user_auth_rm_csv != "n" :
+            print("Merci de répondre par 'Y' pour Oui ou 'n' pour Non.")
+
+if user_auth_rm_csv == "n" :
+    print("Fin de programme !")
+    exit()
 
 #Création du fichier output_scraper.csv et écriture de la ligne d'en-tête
 with open("output_scraper.csv", "w") as csv_file:
@@ -11,7 +27,8 @@ with open("output_scraper.csv", "w") as csv_file:
     header_csv = ["product_page_url", "universal_product_code", "title", "price_including_tax", "price_excluding_tax", "number_available", "product_description", "category", "review_rating", "image_url"]
     writer.writerow(header_csv)
     
-print("Le fichier output_scraper.csv a été créé avec succès !")
+print("Le fichier output_scraper.csv a été initialisé avec succès !")
+print("Extraction de la liste des catégories...")
 
 soup = parser_url.link_to_parse("https://books.toscrape.com/")
